Hard-coding GitHub slug in article template.
This is giving up on the previous approach, because including a
variable that was not rendered resulted in literal `%` in the
content. I looked around to see what sort of "recursive" rendering
is possible (e.g. [1]) but for now will just do this.

I think the best solution would have been to do

    {% if CUSTOM_HEADER_FILENAME %}
      {% include CUSTOM_HEADER_FILENAME %}
    {% endif %}

but that is problematic since the custom file won't be in the
templates directory without some hacking.

[1]: https://stackoverflow.com/a/51104298/1068170